requirements.dev.txt
coverage!=4.3
coveralls
pylama
pytest
pytest-aiohttp
pytest-cov
sqlalchemy_utils
tox
yarl